reverted to original math and found bug in switch code. fixed

+++ firmware/src/display.h
@@ -79,9 +79,14 @@ private:
   uint8_t mode;
 
   // ratio of brightness to frame count for fixed point calculation of LED settings
-  uint8_t secLevelScale;
-  uint8_t minLevelScale;
-  uint8_t hourLevelScale;
+  //uint32_t secLevelScale;
+  //uint32_t minLevelScale;
+  //uint32_t hourLevelScale;
+
+  // ratio of brightness to frame count for calculation of LED settings
+  float secLevelScale;
+  float minLevelScale;
+  float hourLevelScale;
 
   // different effects
   void displayFill(DisplayParams p, uint16_t* dots);
